SpreadsheetApp extends SpreadsheetHistoryAwareStateWidget
- organised methods.
- renamed methods noun first so naming groups.
- history listening methods no longer perform actions only update state.
- improved updating of state within state change methods.
- spreadsheet/SpreadsheetApp.js replaces App.js

- Closes https://github.com/mP1/walkingkooka-spreadsheet-react/issues/873
- App extends SpreadsheetHistoryAwareStateWidget
