+++ src/index.js
@@ -2,13 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {HashRouter} from "react-router-dom";
 import './index.css';
-import App from './App';
+import SpreadsheetApp from './spreadsheet/SpreadsheetApp.js';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
     <React.StrictMode>
         <HashRouter>
-            <App/>
+            <SpreadsheetApp/>
         </HashRouter>
     </React.StrictMode>,
     document.getElementById('root')
